Added brightness changes throughout the day

+++ clock.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 from tm1637 import TM1637
+import datetime
 from time import time, sleep, localtime, strftime
 import RPi.GPIO as GPIO
 import os
@@ -26,14 +27,28 @@ def initialise_db():
   db.commit()
   db.close()
 def remove_alarm(delete):
+    active_days = []
     db = sqlite3.connect(db_file)
     c = db.cursor()
-    c.execute("DELETE FROM alarms WHERE alarm_ID = "+delete)
+    c.execute("SELECT dow FROM alarms WHERE alarm_ID = "+delete)
+    dow = c.fetchone()[0]
+    if dow % 2 != 1: 
+       for i in range(7):
+          if dow & (128 >> i) != 0:
+             active_days.append(weekdays[i])
+       if len(active_days) == 1:
+          c.execute("DELETE FROM alarms WHERE alarm_ID = "+delete)
+       else:
+          active_days.remove (strftime("%A", localtime()))
+          new_dow = 0
+          for i in range(7):
+             if weekdays[i] in active_days:
+                new_dow = new_dow + (128 >> i)
+          c.execute("UPDATE alarms SET dow = "+str(new_dow)+" WHERE alarm_ID = "+delete)
     db.commit()
     db.close()
 def view_db(alarm_time, ID):   
    time = "00:00"
-   repeat = False
    alarm_ID = None
    alarm_set = False
    db = sqlite3.connect(db_file)
@@ -48,11 +63,19 @@ def view_db(alarm_time, ID):
           alarm_set = True
           alarm_time = time
           ID = alarm_ID
-          if dow % 2 == 1:
-              repeat = True
+          #if dow % 2 == 1:
+          #    repeat = True
    db.commit()
    db.close()
-   return (alarm_time, repeat,ID)
+   return (alarm_time,ID)
+def led_brightness():#function to change brightness of the clock during the day
+    d = datetime.datetime.utcnow()
+    if 8 < d.hour < 21:
+        tm.brightness(7)
+    if 21 <= d.hour < 23:
+        tm.brightness(2)
+    elif 8 > d.hour or d.hour >= 23:
+        tm.brightness(0)
 def show_clock(tm):
     t = localtime()
     tm.numbers(t.tm_hour, t.tm_min, False)
@@ -63,7 +86,7 @@ if len(argv) < 2:
 else:
    db_file = sys.argv[1]
 tm = TM1637(CLK, DIO)    
-tm.brightness(0)
+led_brightness()
 radio = False
 source_path="/home/pi/Music/oliver"
 alarm_path="/home/pi/alarms/symlinks"
@@ -71,13 +94,13 @@ flist = []
 for root, dirs, files in os.walk(source_path, followlinks = True):
   for name in files:
     fullname=os.path.join(root,name)
-    if (fnmatch.fnmatch(fullname,'*.mp3') or fnmatch.fnmatch(fullname,'*.flac')):
+    if (fnmatch.fnmatch(fullname,'*.mp3') or fnmatch.fnmatch(fullname,'*.flac') or fnmatch.fnmatch(fullname,'*.m4a')):
       flist.append(fullname)
 alist = []
 for root, dirs, files in os.walk(alarm_path, followlinks = True):
   for name in files:
     fullname=os.path.join(root,name)
-    if (fnmatch.fnmatch(fullname,'*.mp3') or fnmatch.fnmatch(fullname,'*.flac')):
+    if (fnmatch.fnmatch(fullname,'*.mp3') or fnmatch.fnmatch(fullname,'*.flac') or fnmatch.fnmatch(fullname,'*.m4a')):
       alist.append(fullname)
 instance = vlc.Instance()
 player = instance.media_player_new()
@@ -87,9 +110,11 @@ radio.set_media_list(station)
 count = "-1"
 alarm_time = "25:00"
 ID = 0
+normal_volume = 50
+alarm_volume = 20
 alarm_checker = view_db(alarm_time, ID)
 alarm_time = alarm_checker[0] 
-ID = alarm_checker[2]
+ID = alarm_checker[1]
 while True:
     show_clock(tm)
     if GPIO.input(10) == False:
@@ -100,19 +125,23 @@ while True:
        radio.stop()
        tm.show("FLAC")
        player.set_mrl(random.choice(flist))
+       player.audio_set_volume(normal_volume)
        player.play()
     if GPIO.input(11) == False:
        player.stop()
        tm.show("RAD ")
+       player.audio_set_volume(normal_volume)
        radio.play()
     if player.get_state() == vlc.State.Ended:
         player.set_mrl(random.choice(flist))
+        player.audio_set_volume(normal_volume)
         player.play()
     if count != strftime("%M", localtime()):
         alarm_checker = view_db(alarm_time, ID)
         #print ("Next alarm at " + alarm_checker[0])
-        alarm_time = alarm_checker[0] 
-        ID = alarm_checker[2]
+        alarm_time = alarm_checker[0]
+        ID = alarm_checker[1]
+        led_brightness()
         #print (ID)
     if strftime("%H:%M", localtime()) >= alarm_time:
        #print (alarm_time)
@@ -121,11 +150,11 @@ while True:
        radio.stop()
        player.stop()
        player.set_mrl(random.choice(alist))
+       player.audio_set_volume(alarm_volume)
        player.play()
-       if alarm_checker[1] == False: # removes alarm if repeat not set in DB
-          remove_alarm(str(alarm_checker[2]))
+       remove_alarm(str(alarm_checker[1]))
        alarm_checker = view_db(alarm_time, ID)
        alarm_time = alarm_checker[0] 
-       ID = alarm_checker[2]
+       ID = alarm_checker[1]
     count = strftime("%M", localtime())
     sleep(1)
